refactor dbg!
